requirements.txt
async-generator==1.10
attrs==22.2.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
h11==0.14.0
idna==3.4
lxml==4.9.2
outcome==1.2.0
pycparser==2.21
PySocks==1.7.1
requests==2.28.2
selenium==4.7.2
sniffio==1.3.0
sortedcontainers==2.4.0
trio==0.22.0
trio-websocket==0.9.2
undetected-chromedriver==3.2.1
urllib3==1.26.14
websockets==10.4
wsproto==1.2.0
